moved testutils to testutil package
